Pass Custom Interactions for EIP-1271 Signature Validation
